requirements.txt
aioredis==1.3.1
asgi-redis==1.4.3
asgiref==3.1.2
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
beautifulsoup4==4.9.1
boto==2.49.0
boto3==1.14.2
botocore==1.17.2
certifi==2020.4.5.2
cffi==1.14.0
channels==2.4.0
channels-redis==3.1.0
chardet==3.0.4
clarifai==2.6.2
configparser==5.0.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.2.1
cryptography==3.0
daphne==2.5.0
decorator==4.4.2
defusedxml==0.6.0
Django==2.2.6
django-allauth==0.42.0
django-ckeditor==6.0.0
django-common-helpers==0.9.2
django-cors-headers==3.5.0
django-countries==6.1.2
django-crispy-forms==1.9.2
django-cron==0.5.1
django-debug-toolbar==2.2
django-extra-fields==3.0.1
django-js-asset==1.2.2
django-markdown-deux==1.0.5
django-pagedown==2.1.2
django-redis==4.12.1
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
django-storages==1.9.1
django-tinymce==3.1.0
django-tinymce4-lite==1.8.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
EasyProcess==0.3
emoji==0.5.4
future==0.18.2
geographiclib==1.50
geoip2==3.0.0
geopy==2.0.0
googleapis-common-protos==1.52.0
grpcio==1.30.0
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==20.0.1
idna==2.9
incremental==17.5.0
instapy==0.6.9
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
jsmin==2.2.2
jsonschema==2.6.0
markdown2==2.3.9
MarkupSafe==1.1.1
maxminddb==1.5.4
MeaningCloud-python==1.2.2
msgpack==1.0.0
msgpack-python==0.5.6
numpy==1.19.2
oauthlib==3.1.0
openapi-codec==1.3.2
opencv-python==4.4.0.44
Pillow==7.1.2
plyer==1.4.3
protobuf==3.12.2
psycopg2==2.8.6
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-decouple==3.3
python-telegram-bot==12.8
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
regex==2020.7.14
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
selenium==3.141.0
service-identity==18.1.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
stripe==2.49.0
tornado==6.0.4
tqdm==4.48.0
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.25.9
webdriverdownloader==1.1.0.3
zope.interface==5.1.0
